+++ examples/be-example-express-sqlite/server-demo.ts
@@ -43,9 +43,7 @@ const main = async () => {
                 agent: new AgentEcho(),
                 metadata: {
                     displayName: "Echo Agent",
-                    primer: "This is demo echo agent. Write me a message, and I will send it back to you!",
-                    humanRole: "Human",
-                    aiRole: "AI"
+                    primer: "This is demo echo agent. Write me a message, and I will send it back to you!"
                 }
             }
         }, "default")