+++ packages/convostack-auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @convostack/auth
 
+## 0.0.5
+
+### Patch Changes
+
+- Standardize message roles
+- Updated dependencies
+  - @convostack/models@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes