+++ packages/convostack/CHANGELOG.md
@@ -1,5 +1,33 @@
 # convostack
 
+## 0.0.6
+
+### Patch Changes
+
+- Standardize message roles
+- Updated dependencies
+  - @convostack/langchain-memory@0.0.6
+  - @convostack/shared@0.0.5
+  - @convostack/agent@0.0.5
+  - config-tsconfig@0.0.3
+  - @convostack/agent-echo@0.0.5
+  - @convostack/agent-sse@0.0.5
+  - @convostack/agent-sse-echo-server@0.0.5
+  - @convostack/auth@0.0.5
+  - @convostack/auth-jwt@0.0.5
+  - @convostack/backend-express@0.0.5
+  - @convostack/frontend-react@0.0.5
+  - @convostack/models@0.0.5
+  - @convostack/storage-engine-prisma-sqlite@0.0.5
+
+## 0.0.5
+
+### Patch Changes
+
+- Add langchain memory
+- Updated dependencies
+  - @convostack/langchain-memory@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes